Move the fixture out to conftest.py
